int para double

+++ Program.cs
@@ -1,10 +1,10 @@
-﻿int v,d,t;
+﻿double v,d,t;
 
 Console.Write("Distância percorrida (m): ");
-d = Convert.ToInt32(Console.ReadLine());
+d = Convert.ToDouble(Console.ReadLine());
 
 Console.Write("Tempo gasto (s): ");
-t = Convert.ToInt32(Console.ReadLine());
+t = Convert.ToDouble(Console.ReadLine());
 
 v = d / t;
 Console.Write($"Velocidade média: {v} m/s ");
